refactor attrs tests
